+++ src/main/java/com/intellij/plugins/haxe/lang/psi/fakes/HaxeFakeNamedComponent.java
@@ -1,21 +1,24 @@
 package com.intellij.plugins.haxe.lang.psi.fakes;
 
 import com.intellij.plugins.haxe.HaxeComponentType;
-import com.intellij.plugins.haxe.lang.psi.HaxeComponentName;
 import com.intellij.plugins.haxe.lang.psi.HaxeNamedComponent;
 import com.intellij.plugins.haxe.metadata.HaxeMetadataList;
 import com.intellij.plugins.haxe.metadata.psi.HaxeMeta;
 import com.intellij.plugins.haxe.metadata.psi.impl.HaxeMetadataTypeName;
+import com.intellij.plugins.haxe.model.HaxeMethodModel;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.meta.PsiMetaData;
 import com.intellij.psi.tree.IElementType;
+import org.jetbrains.annotations.Nls;
+import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+
+import javax.swing.*;
 
 public abstract class HaxeFakeNamedComponent  extends HaxeFakePsiElement implements HaxeNamedComponent {
 
     private static final HaxeMetadataList EMPTY_META = new HaxeMetadataList();
-
-    public abstract HaxeComponentType componentType();
 
     @Override
     public @Nullable HaxeNamedComponent getTypeComponent() {
@@ -63,9 +66,8 @@ public abstract class HaxeFakeNamedComponent  extends HaxeFakePsiElement impleme
     }
 
     @Override
-    public HaxeComponentType getComponentType() {
-        return null;
-    }
+    @NotNull
+    public abstract HaxeComponentType getComponentType();
 
     @Override
     public IElementType getTokenType() {
@@ -73,9 +75,18 @@ public abstract class HaxeFakeNamedComponent  extends HaxeFakePsiElement impleme
     }
 
     @Override
-    public String getDocs() {
+    public String getDocsText() {
         return "";
     }
+
+    @Override
+    public HaxeNamedComponent getDocsPsi() {
+        return null;
+    }
+
+    @Override
+    public abstract String getName();
+
 
     @Override
     public @NotNull HaxeMetadataList getMetadataList(@Nullable Class<? extends HaxeMeta> metadataType) {
@@ -87,4 +98,10 @@ public abstract class HaxeFakeNamedComponent  extends HaxeFakePsiElement impleme
         return false;
     }
 
+    @Nullable
+    public HaxeMethodModel getModel() {
+        return null;
+    }
+
+    public abstract String getQuickNavigateInfo();
 }