+++ src/main/java/com/intellij/plugins/haxe/model/evaluator/HaxeExpressionEvaluatorHandlers.java
@@ -650,7 +650,7 @@ public class HaxeExpressionEvaluatorHandlers {
     // @TODO: check if it has string interpolation inside, in that case text is not constant
     String constant = HaxeTypeLiteralsUtils.translateHaxeStringToJavaString(element.getText());
     constant = constant.substring(1, constant.length()-1); //drop the quotes wrapping the string
-    return SpecificHaxeClassReference.primitive("String", element, constant).createHolder();
+    return SpecificHaxeClassReference.getString(element, constant).createHolder();
   }
 
   static ResultHolder handleSwitchCaseCaptureVar(HaxeGenericResolver resolver, HaxeSwitchCaseCaptureVar captureVar) {
@@ -1715,6 +1715,11 @@ public class HaxeExpressionEvaluatorHandlers {
     ResultHolder assignHint = resolver.getAssignHint();
     SpecificTypeReference suggestedKeyType = null;
     SpecificTypeReference suggestedValueType = null;
+    // checkEnumMemberHints in HaxeResolver will try to detectType of refrences and in resolve paths it triggers
+    // handleMapLiteral without assignHint, and this can result in an EnumType when EnumValue is expected
+    if(assignHint == null && mapLiteral.getParent() instanceof HaxeVarInit init) {
+      assignHint = tryGetHintFromFieldTypeTag(init);
+    }
     if (assignHint != null) {
       SpecificHaxeClassReference hintClassType = assignHint.getClassType();
       if (hintClassType != null) {
@@ -2772,21 +2777,6 @@ public class HaxeExpressionEvaluatorHandlers {
     return null;
   }
 
-  //private static boolean containsTypeParameters(ResultHolder holder) {
-  //  if (holder.isUnknown()) return  false;
-  //  if (holder.isTypeParameter()) return true;
-  //  SpecificTypeReference type = holder.getType();
-  //  if (type instanceof  SpecificHaxeClassReference classReference) {
-  //    for (ResultHolder specific : classReference.getSpecifics()) {
-  //      if (containsTypeParameters(specific)) return  true;
-  //    }
-  //  }
-  //  if (type instanceof SpecificFunctionReference  function) {
-  //    if(!function.getTypeParameters().isEmpty()) return true;
-  //  }
-  //  return false;
-  //}
-
   static boolean isUntypedReturn(HaxeReturnStatement statement) {
     PsiElement child = statement.getFirstChild();
     while(child != null) {
@@ -2799,6 +2789,18 @@ public class HaxeExpressionEvaluatorHandlers {
   }
 
 
+
+  private static ResultHolder tryGetHintFromFieldTypeTag(HaxeVarInit init) {
+    HaxePsiField field = PsiTreeUtil.getParentOfType(init, HaxePsiField.class);
+    if(field != null){
+      HaxeTypeTag typeTag = field.getTypeTag();
+      if(typeTag != null)  {
+        ResultHolder type = HaxeTypeResolver.getTypeFromTypeTag(typeTag, field);
+        if(type != null && !type.isUnknown()) return type;
+      }
+    }
+    return null;
+  }
 
 
 }