+++ src/test/java/com/intellij/plugins/haxe/lang/parser/expressions/ExpressionTest.java
@@ -173,4 +173,14 @@ public class ExpressionTest extends HaxeParsingTestBase {
   public void testInlineCallExpression() throws Throwable {
     doTest(true);
   }
+
+  @Test
+  public void testRegularExpressions() throws Throwable {
+    doTest(true, true);
+  }
+
+  @Test
+  public void testXmlLiterals() throws Throwable {
+    doTest(true);
+  }
 }